Refactor: Create `BaseRecaptchaEnterpriseCallbackTest` to reduce code duplication
Introduces an abstract base class, `BaseRecaptchaEnterpriseCallbackTest`, to encapsulate common setup logic for reCAPTCHA Enterprise callback tests. This includes initializing the `Journey` instance and setting the test tree configuration.

`RecaptchaEnterpriseCallbackTest` and `RecaptchaEnterpriseCallbackFailureTest` are updated to inherit from this new base class, removing redundant setup code and simplifying their implementation.
